A sketch of the next assignments
The originally envisioned asst 009 and 010 are too large on their own. They need to be broken down into about 5-6 assignments, based on the expanding on the major points in them and extending the hands-on to achieve a smooth progression with about the same slope (logarithmic, as it is expected from the human mind) as asst 001-008.

Asst 001-005 have to be redone in the new format.